pyproject.toml
[project]
name = "eleventh3"
version = "0.1.0"
description = "Web3 implementation of Eleventh."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "vyper (>=0.4.3,<0.5.0)",
    "titanoboa (>=0.2.8,<0.3.0)",
    "pytest-bdd (>=8.1.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
